* 'pedestrian_crossing' of https://github.com/abhi2039/valhalla:
  build: Bail with error if non-existant build-type specified (#2965)
  nit: Enables compiler warnings in part of mjolnir module (#2922)
  fix missing comma scripts/valhalla_build_config (#2963)
  Remove duplicate names being used across different gurka tests (#2962)
  Dont abort bidirectional a* search if only one direction is exhausted (#2936)
  penalize uturns at pencil points and short internal edges (#2944)
  Remove unused IsNextEdgeInternal function (#2958)
  restore compatibility with gcc 6.3.0, libprotobuf 3.0.0, boost v1.62.0 (#2953)
  Add ability to build Valhalla modules as STATIC libraries (#2957)
  Allow disabling Werror (#2937)
  Enhanced logic for IsTurnChannelManeuverCombinable (#2952)
  Allow config object to be passed-in to path algorithms (#2949)
  Floating-point precision fix in meili::Project() fixes map-match issue (#2946)
  Use kServiceRoad edges while reclassifying ferry connections (#2933)
  ci: Update cache key to trigger clearing it (#2947)

+++ src/meili/candidate_search.cc
@@ -67,9 +67,9 @@ CandidateCollector::WithinSquaredDistance(const midgard::PointLL& location,
 
     // Projection information
     midgard::PointLL point;
-    float sq_distance = 0.f;
+    double sq_distance = 0.0;
     size_t segment;
-    float offset;
+    double offset;
 
     baldr::GraphId snapped_node;
     baldr::PathLocation correlated(baldr::Location(location, stop_type));
@@ -81,10 +81,10 @@ CandidateCollector::WithinSquaredDistance(const midgard::PointLL& location,
       std::tie(point, sq_distance, segment, offset) = helpers::Project(projector, shape);
 
       if (sq_distance <= sq_search_radius) {
-        const float dist = edge->forward() ? offset : 1.f - offset;
-        if (dist == 1.f) {
+        const double dist = edge->forward() ? offset : 1.0 - offset;
+        if (dist == 1.0) {
           snapped_node = edge->endnode();
-        } else if (dist == 0.f) {
+        } else if (dist == 0.0) {
           snapped_node = opp_edge->endnode();
         }
         correlated.edges.emplace_back(edgeid, dist, point, sq_distance);
@@ -100,10 +100,10 @@ CandidateCollector::WithinSquaredDistance(const midgard::PointLL& location,
         std::tie(point, sq_distance, segment, offset) = helpers::Project(projector, shape);
       }
       if (sq_distance <= sq_search_radius) {
-        const float dist = opp_edge->forward() ? offset : 1.f - offset;
-        if (dist == 1.f) {
+        const double dist = opp_edge->forward() ? offset : 1.0 - offset;
+        if (dist == 1.0) {
           snapped_node = opp_edge->endnode();
-        } else if (dist == 0.f) {
+        } else if (dist == 0.0) {
           snapped_node = edge->endnode();
         }
         correlated.edges.emplace_back(opp_edgeid, dist, point, sq_distance);
